fixing red hat

+++ install.sh
@@ -26,7 +26,6 @@ then
     DISTRO=$(cat /etc/os-release | grep -w ID_LIKE | cut -d= -f2 | tr -d '=')
 
     echo -e ${YELLOW}"[*] ${GREEN}Found:${BOLD_GREEN}" ${OS_NAME} ${GREEN}"Version: "${BOLD_GREEN}${OS_VERSION}
-    echo -e ${YELLOW}"[*] ${CYAN}Checking python version installed currently on the system..."${RESET}
 
     # Add Python on RedHat 7
     if [ "$OS_NAME" == *"Red Hat"* ];
@@ -51,10 +50,16 @@ then
     # Add Python PPA on Ubuntu
     if [[ "$OS_NAME" == *"Ubuntu"* ]];
     then     
-        add-apt-repository ppa:fkrull/deadsnakes
-        apt-get update
-        apt-get install python${PYSETENV_PYTHON_VERSION}
-        apt-get autoremove -y
+        if [ hash python${PYSETENV_PYTHON_VERSION}];
+        then
+            echo -e ${YELLOW}"[*] ${CYAN}Checking python version installed currently on the system..."${RESET}
+            echo -e ${YELLOW}"[*] " ${BOLD_GREEN}Found $(python -V)${RESET}
+        else
+            add-apt-repository ppa:fkrull/deadsnakes
+            apt-get update
+            apt-get install python${PYSETENV_PYTHON_VERSION}
+            apt-get autoremove -y
+        fi
     fi
 else
     # Add Python on CentOS
